Cleans up common, core and full after the code split
