+++ core/src/main/java/apoc/CoreApocGlobalComponents.java
