+++ full/src/main/java/apoc/generate/BaseGraphGenerator.java
