+++ full/src/main/java/apoc/generate/config/BasicGeneratorConfig.java
