+++ full/src/main/java/apoc/generate/distribution/MutableSimpleDegreeDistribution.java
