+++ full/src/main/java/apoc/generate/distribution/SimpleDegreeDistribution.java
