+++ full/src/main/java/apoc/generate/relationship/CompleteGraphRelationshipGenerator.java
