+++ full/src/test/java/apoc/generate/config/BarabasiAlbertConfigTest.java
