+++ full/src/test/java/apoc/generate/config/WattsStrogatzConfigTest.java
